III-7336 migrate Button component to shadcn
